* 1785_editor_normal_color:
  Some optimization and cleanup of color-related code in skin layer.
  Some optimization and cleanup of color-related code in TTY layer.
  Ticket #1785: fix of editor default color usage.

+++ lib/tty/color-internal.h
@@ -25,7 +25,7 @@ typedef struct mc_color_pair_struct
     const char *cbg;
     int ifg;
     int ibg;
-    size_t pair_index;
+    int pair_index;
     gboolean is_temp;
 } tty_color_pair_t;
